feat(FIR-21051): Add expected_output_is_regex to RunTestCaseResultBase (#5)
This commit introduces a new data member called expected_output_is_regex_ to
RunTestCaseResultBase. If set to true, the expected output in a SQL test is treated as
a RE2 regex that must match the entire actual output. Otherwise, the current behavior is
used where the expected output must the actual output literally.  If alternation groups
are used, each group is matched separately.